+++ src/LanguageServer.cpp
@@ -920,7 +920,7 @@ void LanguageServer::onDidChangeWorkspaceFolders(const lsp::DidChangeWorkspaceFo
 
 void LanguageServer::onDidChangeWatchedFiles(const lsp::DidChangeWatchedFilesParams& params)
 {
-    Luau::DenseHashMap<WorkspaceFolderPtr, std::vector<lsp::FileEvent>> workspaceChanges{nullptr};
+    Luau::DenseHashMap2<WorkspaceFolderPtr, std::vector<lsp::FileEvent>> workspaceChanges;
 
     for (const auto& change : params.changes)
     {