+++ src/platform/overdare/OverdareLuauExt.cpp
@@ -118,7 +118,7 @@ std::optional<Luau::WithPredicate<Luau::TypePackId>> MagicInstanceIsA::handleOld
 
     auto type = Luau::follow(tfun->type);
 
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+    Luau::TypeArena& arena = *typeChecker.currentModule->internalTypes;
     Luau::TypePackId booleanPack = arena.addTypePack({typeChecker.booleanType});
     return Luau::WithPredicate<Luau::TypePackId>{booleanPack, {Luau::IsAPredicate{std::move(*lvalue), expr.location, type}}};
 }
@@ -136,7 +136,7 @@ bool MagicInstanceIsA::infer(const Luau::MagicFunctionCallContext& context)
     std::string className(str->value.data, str->value.size);
     std::optional<Luau::TypeFun> tfun = context.solver->rootScope->lookupType(className);
     if (!tfun)
-        context.solver->reportError(
+        context.solver->DEPRECATED_reportError(
             Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}});
 
     return false;
@@ -180,7 +180,7 @@ std::optional<Luau::WithPredicate<Luau::TypePackId>> MagicInstanceClone::handleO
     if (!index)
         return std::nullopt;
 
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+    Luau::TypeArena& arena = *typeChecker.currentModule->internalTypes;
     auto instanceType = typeChecker.checkExpr(scope, *index->expr);
     return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({instanceType.type})};
 }
@@ -214,7 +214,7 @@ std::optional<Luau::WithPredicate<Luau::TypePackId>> MagicInstanceFromExisting::
     if (expr.args.size < 1)
         return std::nullopt;
 
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+    Luau::TypeArena& arena = *typeChecker.currentModule->internalTypes;
     auto instanceType = typeChecker.checkExpr(scope, *expr.args.data[0]);
     return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({instanceType.type})};
 }
@@ -262,7 +262,7 @@ std::optional<Luau::WithPredicate<Luau::TypePackId>> MagicInstanceFindFirstXWhic
 
     auto type = Luau::follow(tfun->type);
 
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+    Luau::TypeArena& arena = *typeChecker.currentModule->internalTypes;
     Luau::TypeId nillableClass = Luau::makeOption(typeChecker.builtinTypes, arena, type);
     return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({nillableClass})};
 }
@@ -280,7 +280,7 @@ bool MagicInstanceFindFirstXWhichIsA::infer(const Luau::MagicFunctionCallContext
     std::optional<Luau::TypeFun> tfun = context.solver->rootScope->lookupType(className);
     if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
     {
-        context.solver->reportError(Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}, str->location);
+        context.solver->DEPRECATED_reportError(Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}, str->location);
         return false;
     }
 
@@ -326,7 +326,7 @@ std::optional<Luau::WithPredicate<Luau::TypePackId>> MagicEnumItemIsA::handleOld
 
     auto type = Luau::follow(tfun->type);
 
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+    Luau::TypeArena& arena = *typeChecker.currentModule->internalTypes;
     Luau::TypePackId booleanPack = arena.addTypePack({typeChecker.booleanType});
     return Luau::WithPredicate<Luau::TypePackId>{booleanPack, {Luau::IsAPredicate{std::move(*lvalue), expr.location, type}}};
 }
@@ -344,7 +344,7 @@ bool MagicEnumItemIsA::infer(const Luau::MagicFunctionCallContext& context)
     std::string enumItem(str->value.data, str->value.size);
     std::optional<Luau::TypeFun> tfun = context.constraint->scope->lookupImportedType("Enum", enumItem);
     if (!tfun)
-        context.solver->reportError(
+        context.solver->DEPRECATED_reportError(
             Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownSymbol{enumItem, Luau::UnknownSymbol::Type}});
 
     return false;
@@ -433,7 +433,7 @@ bool MagicInstancePropertyCheck::infer(const Luau::MagicFunctionCallContext& con
     std::string property(str->value.data, str->value.size);
     if (!Luau::lookupExternTypeProp(ctv, property))
     {
-        context.solver->reportError(Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownProperty{*selfTy, property}});
+        context.solver->DEPRECATED_reportError(Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownProperty{*selfTy, property}});
     }
 
     return false;
@@ -476,7 +476,7 @@ std::optional<Luau::WithPredicate<Luau::TypePackId>> MagicQueryDescendants::hand
         classTypes.push_back(Luau::follow(tfun->type));
     }
 
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+    Luau::TypeArena& arena = *typeChecker.currentModule->internalTypes;
 
     Luau::TypeId elementType;
     if (classTypes.size() == 1)
@@ -513,7 +513,7 @@ bool MagicQueryDescendants::infer(const Luau::MagicFunctionCallContext& context)
         std::optional<Luau::TypeFun> tfun = context.solver->rootScope->lookupType(className);
         if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
         {
-            context.solver->reportError(Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}, str->location);
+            context.solver->DEPRECATED_reportError(Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}, str->location);
             return false;
         }
         classTypes.push_back(Luau::follow(tfun->type));
@@ -593,7 +593,7 @@ std::optional<Luau::WithPredicate<Luau::TypePackId>> MagicTypeLookup::handleOldS
 
             auto type = Luau::follow(tfun->type);
 
-            Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+            Luau::TypeArena& arena = *typeChecker.currentModule->internalTypes;
             Luau::TypePackId classTypePack = arena.addTypePack({type});
             return Luau::WithPredicate<Luau::TypePackId>{classTypePack};
         }
@@ -619,7 +619,7 @@ bool MagicTypeLookup::infer(const Luau::MagicFunctionCallContext& context)
             std::optional<Luau::TypeFun> tfun = context.solver->rootScope->lookupType(className);
             if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
             {
-                context.solver->reportError(
+                context.solver->DEPRECATED_reportError(
                     Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}});
                 return false;
             }
@@ -631,7 +631,7 @@ bool MagicTypeLookup::infer(const Luau::MagicFunctionCallContext& context)
         }
         else
         {
-            context.solver->reportError(
+            context.solver->DEPRECATED_reportError(
                 Luau::TypeError{context.callSite->args.data[0]->location, Luau::GenericError{errorMessagePrefix + " '" + className + "'"}});
         }
     }